Move adaptive_theme
