+++ example/lib/shared/const/app.dart
@@ -4,8 +4,8 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 import '../controllers/theme_controller.dart';
+import '../model/adaptive_theme.dart';
 import '../widgets/app/responsive_scaffold.dart';
-import 'adaptive_theme.dart';
 
 // ignore_for_file: comment_references
