+++ example/lib/shared/model/adaptive_theme.dart
